// angular-components-oasis/projects/oasis-button/src/shared/button-type.enum.ts
/**
 * Типы кнопок
 */
export enum ButtonType {
  /** Основной тип кнопок */
  Main = 'main',

  /** Основной тип кнопок с прозрачным фоном */
  MainTransparent = 'main-transparent',

  /** Кнопка без рамки */
  WithoutBorder = 'without-border',

  /** Ссылка */
  Link = 'link'
}