// angular-components-oasis/projects/oasis-button/src/shared/button-state.enum.ts
/**
 * Состояние кнопки
 */
export enum ButtonState {
  /** Стандартное состояние */
  Default = 'default',

  /** Наведенное состояние */
  Hover = 'hover',

  /** Нажатое состояние */
  Click = 'click',

  /** Выключенное состояние */
  Disabled = 'disabled'
}